fix: fix time parsing issue
